fix: move fs.js to the node folder
