+++ packages/puppeteer-core/src/node/ChromeLauncher.ts
@@ -25,7 +25,6 @@ import {
 
 import {debugError} from '../common/util.js';
 import {assert} from '../util/assert.js';
-import {rm} from '../util/fs.js';
 
 import {
   BrowserLaunchArgumentOptions,
@@ -34,6 +33,7 @@ import {
 } from './LaunchOptions.js';
 import {ProductLauncher, ResolvedLaunchArgs} from './ProductLauncher.js';
 import {PuppeteerNode} from './PuppeteerNode.js';
+import {rm} from './util/fs.js';
 
 /**
  * @internal
